% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koweps-package.R
\docType{package}
\name{koweps-package}
\alias{koweps}
\alias{koweps-package}
\title{koweps: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
